Inicializacion de datos y correciones (#7)
Puse los initializeData() en los repositories.
Corregi la logica de estadisticas que tenia hardcodeado los datos para que use los datos generados de los repos
Agrego readme con los endpoints

+++ src/main/java/org/utn/ba/tptacsg2/dtos/EstadisticasUso.java
@@ -4,19 +4,19 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record EstadisticasUso(
     @JsonProperty("cantidad_eventos")
-    Long cantidadEventos,
+    Integer cantidadEventos,
 
     @JsonProperty("cantidad_eventos_activos")
-    Long cantidadEventosActivos,
+    Integer cantidadEventosActivos,
 
     @JsonProperty("cantidad_inscripciones_totales")
-    Long cantidadInscripcionesTotales,
+    Integer cantidadInscripcionesTotales,
 
     @JsonProperty("cantidad_inscripciones_confirmadas")
-    Long cantidadInscripcionesConfirmadas,
+    Integer cantidadInscripcionesConfirmadas,
 
     @JsonProperty("cantidad_inscripciones_waitlist")
-    Long cantidadInscripcionesWaitlist,
+    Integer cantidadInscripcionesWaitlist,
 
     @JsonProperty("tasa_conversion_waitlist")
     Double tasaConversionWaitlist,
